# src/main/webapp/docs/swagger.yaml
openapi: 3.0.0
servers: 
  - url: 'https://api-electronicdevices-313314.nw.r.appspot.com/api/v1/'
  - url: 'http://localhost:8089/api/v1/'
info:
  description: This API can be used to consult multiple electronic devices which could be found in different shops.
  version: 2.0.0
  title: API-ElectronicDevices
  contact:
    email: joselugarciamarin2406@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
  /devices:
    get:
      tags:
        - Devices
      summary: Get devices
      description: Get all the devices
      operationId: getDevices
      parameters:
        - name: order
          in: query
          description: Order results by name, price, screenSize, weight, memory or storage.
          required: false
          schema:
            type: string
            enum: 
              - name
              - '-name'
              - price
              - '-price'
              - screenSize
              - '-screenSize'
              - weight
              - '-weight'
              - memory
              - '-memory'
              - storage
              - '-storage'
        - name: brand
          in: query
          description: Query to filter by brand
          required: false
          schema:
            type: string
        - name: deviceType
          in: query
          description: Query to filter by device type
          required: false
          schema:
            type: string
        - name: Color
          in: query
          description: Query to filter by color
          required: false
          schema:
            type: string
        - name: OS
          in: query
          description: Query to filter by OS
          required: false
          schema:
            type: string
        - name: offset
          in: query
          description: Offset of the results
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: limit
          in: query
          description: Maximum number of results
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '200':
          description: Search resutl matching criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Not found
    post:
      tags:
        - Devices
      summary: Create a device
      description: Create a device
      operationId: createDevice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceEmpty'
        description: Device to create
        required: true
      responses:
        '201':
          description: Device created
        '400':
          description: Bad request
    put:
      tags:
        - Devices
      summary: Update a device
      description: Update a device
      operationId: updateDevice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Device'
      responses:
        '204':
          description: Device update
        '400':
          description: Device not update
...
  /complements:
    get:
      tags:
        - Complements
      summary: Get complements
      description: Get all the complements
      operationId: getComplements
      parameters:
        - name: order
          in: query
          description: Order results by name or price
          required: false
          schema:
            type: string
            enum:
              - name
              - '-name'
              - price
              - '-price'
        - name: isEmpty
          in: query
          description: Filter complements with no devices
          required: false
          schema:
            type: boolean
        - name: name
          in: query
          description: Filter complements by name
          required: false
          schema: 
            type: string
        - name: material
          in: query
          description: Filter complements by material
          schema:
            type: string
      responses:
        '200':
          description: Search results matching criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Complement'
        '400':
          description: Bad request
    post:
      tags:
        - Complements
      summary: Create a complement
      description: Create a complement
      operationId: createComplement
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplementEmpty'
        description: Device to create
        required: true
      responses:
        '201':
          description: Complement created
        '400':
          description: Bad reeques
    put:
      tags:
        - Complements
      summary: Update a complement
      description: Upadate a complement
      operationId: updateComplement
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Complement'
        description: Complement to update
        required: true
      responses:
        '204':
          description: Complement updated
        '400':
          description: Bad request
  '/complements/{id}':
    get:
      tags:
        - Complements
      summary: Get a complement by id
      description: Get a complement  by id
      operationId: getComplement
      parameters:
        - name: id
          in: path
          description: Id of the complement
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses: 
        '200':
          description: Complement found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Complement'
        '400':
          description: Complement not found
    delete:
      tags:
        - Complements
      summary: Delete a complement
      description: Delete a complement
      operationId: deleteComplement
      parameters:
        - name: id
          in: path
          description: Id of the complement
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '204':
          description: Complement deleted
        '404':
          description: Complement not found
  '/complements/{complementId}/{deviceId}':
    post:
      tags: 
        - Complements
      summary: Add a device compatible with a complemet
      description: Add a device compatible with a complement
      operationId: addDeviceToComplement
      requestBody:
        content:
          text/plain:
            schema:
              type: string
      parameters: 
        - name: complementId
          in: path
          description: Id of the complement
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: deviceId
          in: path
          description: Id of the device
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '201':
          description: Device added to the compatible devices list with a complement
        '400':
          description: Bad request
        '404':
          description: Accessory or device not found
    delete:
      tags: 
        - Complements
      summary: Delete a device compatible with a complement
      description: Delete a device compatible with a complement
      operationId: deleteDeviceFromComplement
      parameters: 
        - name: complementId
          in: path
          description: Id of the complement
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: deviceId
          in: path
          description: Id of the device
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '204':
          description: Device added to the compatible devices list with a complement 
        '404':
          description: Accessory or device not found
  /accessories:
    get:
      tags:
        - Accessories
      summary: Get accessories
      description: Get all the accessories
      operationId: getAccesories
      parameters:
        - name: order
          in: query
          description: Order results by name or price
          required: false
          schema:
            type: string
            enum:
              - name
              - '-name'
              - price
              - '-price'
        - name: isEmpty
          in: query
          description: Filter accessories with no devices
          required: false
          schema:
            type: boolean
        - name: name
          in: query
          description: Filter accessories by name
          required: false
          schema: 
            type: string
        - name: wireless
          in: query
          description: Filter accessories by wireless
          schema:
            type: string
            enum:
            - 'true'
            - 'false'
      responses:
        '200':
          description: Search results matching criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accessory'
        '400':
          description: Bad request
    post:
      tags:
        - Accessories
      summary: Create a accessory
      description: Create a accessory
      operationId: createAccessory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessoryEmpty'
        description: Accessory to create
        required: true
      responses:
        '201':
          description: Accessory created
        '400':
          description: Bad request
    put:
      tags:
        - Accessories
      summary: Update an accessory
      description: Upadate an accessory
      operationId: updateAccessory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Accessory'
        description: Accessory to update
        required: true
      responses:
        '204':
          description: Accessory updated
        '400':
          description: Bad request
  '/accessories/{accessoryId}/{deviceId}':
    post:
      tags: 
        - Accessories
      summary: Add a device compatible with an accessory
      description: Add a device compatible with an accessory
      operationId: addDeviceToAccessories
      requestBody:
        content:
          text/plain:
            schema:
              type: string
      parameters: 
        - name: accessoryId
          in: path
          description: Id of the accessory
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: deviceId
          in: path
          description: Id of the device
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '201':
          description: Device added to the compatible devices list with an accesory
        '400':
          description: Bad request
        '404':
          description: Accessory or device not found
    delete:
      tags: 
        - Accessories
      summary: Delete a device compatible with an accessory
      description: Delete a device compatible with an accessory
      operationId: deleteDeviceFromAccesory
      parameters: 
        - name: accessoryId
          in: path
          description: Id of the accessory
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: deviceId
          in: path
          description: Id of the device
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '204':
          description: Device added to the compatible devices list with an accesory 
        '404':
          description: Accessory or device not found
  '/accessories/{id}':
    get:
      tags:
        - Accessories
      summary: Get an accessory by id
      description: Get an accessory  by id
      operationId: getAccessory
      parameters:
        - name: id
          in: path
          description: Id of the complement
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses: 
        '200':
          description: Accessory found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accessory'
        '400':
          description: Accessory not found
    delete:
      tags:
        - Accessories
      summary: Delete an accessory
      description: Delete an accessory
      operationId: deleteAccessory
      parameters:
        - name: id
          in: path
          description: Id of the complement
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '204':
          description: Accessory deleted
        '404':
          description: Accessory not found
  /shops:
    get:
      tags: 
      - Shops
      summary: Get all the shops
      description: Get all the shops
      operationId: getAllShops
      parameters:
        - name: order
          in: query
          description: Order result by name
          required: false
          schema:
            type: string
            enum: 
              - name
              - '-name'
        - name: isEmpty
          in: query
          description: Filter shops with no devices
          required: false
          schema:
            type: boolean
        - name: name
          in: query
          description: Filter shops by name
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Search results matching criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shop'
        '400':
          description: Bad request
    post:
      tags:
        - Shops
      summary: Create a shop
      description: Create a shop
      operationId: createShop
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopEmpty'
        required: true
      responses:
        '201':
          description: Shop created
        '400':
          description: Bad request
    put:
      tags: 
        - Shops
      summary: Update a shop
      description: Update a shop
      operationId: updateShop
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopEmpty'
      responses:
        '204':
          description: Shop update
        '400':
          description: Bad request
        '404':
          description: Shop not found
  '/shops/{id}':
    get:
      tags:
        - Shops
      summary: Get a shop by id
      description: Get a shop by id
      operationId: getShopById
      parameters: 
        - name: id
          in: path
          description: Id of the shop
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '204':
          description: Shop found
        '404':
          description: Shop not found
    delete:
      tags: 
        - Shops
      summary: Delete a shop
      description: Delete a shop
      operationId: deleteShop
      parameters: 
        - name: id
          in: path
          description: Id of the shop
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '204':
          description: Shop delete
        '404':
          description: Shop not found
  '/shops/{shopId}/{deviceId}':
    post:
      tags: 
        - Shops
      summary: Add device to shop
      description: Add device to shop
      operationId: addDeviceToShop
      requestBody:
        content:
          text/plain:
            schema:
              type: string
      parameters: 
        - name: shopId
          in: path
          description: Id of the shop
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: deviceId
          in: path
          description: Id of the device
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '201':
          description: Device added to the shop
        '400':
          description: Bad request
        '404':
          description: Shop or device not found
    delete:
      tags: 
        - Shops
      summary: Delete a device from a shop
      description: Delete a device from a shop
      operationId: deleteDeviceFromShop
      parameters: 
        - name: shopId
          in: path
          description: Id of the shop
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: deviceId
          in: path
          description: Id of the device
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '204':
          description: Device delete from shop
        '404':
          description: Shop or device not found
  '/shops/complements/{shopId}/{complementId}':
    post:
      tags: 
        - Shops
      summary: Add complement to shop
      description: Add complement to shop
      operationId: addComplementToShop
      requestBody:
        content:
          text/plain:
            schema:
              type: string
      parameters: 
        - name: shopId
          in: path
          description: Id of the shop
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: complementId
          in: path
          description: Id of the complement
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '201':
          description: Complement added to the shop
        '400':
          description: Bad request
        '404':
          description: Shop or complement not found
    delete:
      tags: 
        - Shops
      summary: Delete a complement from a shop
      description: Delete a complement from a shop
      operationId: deleteComplementFromShop
      parameters: 
        - name: shopId
          in: path
          description: Id of the shop
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: complementId
          in: path
          description: Id of the complement
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '204':
          description: Complement delete from shop
        '404':
          description: Shop or complement not found
  '/shops/accessories/{shopId}/{accessoryId}': 
    post:
      tags: 
        - Shops
      summary: Add accessory to shop
      description: Add accessory to shop
      operationId: addAccessoryToShop
      requestBody:
        content:
          text/plain:
            schema:
              type: string
      parameters: 
        - name: shopId
          in: path
          description: Id of the shop
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: accessoryId
          in: path
          description: Id of the accessory
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '201':
          description: Accessory added to the shop
        '400':
          description: Bad request
        '404':
          description: Shop or accessory not found
    delete:
      tags: 
        - Shops
      summary: Delete an accessory from a shop
      description: Delete an accessory from a shop
      operationId: deleteAccessoryFromShop
      parameters: 
        - name: shopId
          in: path
          description: Id of the shop
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: accessoryId
          in: path
          description: Id of the accessory
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '204':
          description: Accessory delete from shop
        '404':
          description: Shop or accessory not found
components:
  schemas:
    Device:
      type: array
      items:
        type: object
        required: 
          - name
        properties:
          id:
            type: integer
          name:
            type: string
          brand:
            type: string
          price: 
            type: number
          screenSize:
            type: number
          weight:
            type: number
          deviceType:
            type: string
          storage:
            type: integer
          processor:
            type: string
          memory:
            type: integer
          OS:
            type: string
          setColors:
            type: array
            items:
              type: string
          camera:
            type: string
      example:
        id: 1
        name: Galaxy s20 FE
        brand: Samsung
        price: 649.0
        screenSize: 6.5
        weight: 0.19
        deviceType: MOBILE
        storage: 128
        processor: Samsung Exynos 990
        memory: 6
        setColors: [Cloud Orange, Cloud Red, Cloud white, Cloud Navy]
        camera: "Principal 12Mp"
        os: Android
    DeviceEmpty:
      type: array
      items:
        type: object
        required: 
          - name
        properties:
          id:
            type: integer
          name:
            type: string
          brand:
            type: string
          price: 
            type: number
          screenSize:
            type: number
          weight:
            type: number
          deviceType:
            type: string
          storage:
            type: integer
          processor:
            type: string
          memory:
            type: integer
          setColors:
            type: array
            items:
              type: string
          OS:
            type: string
          camera:
            type: string
      example:
        id: 1
        name: Galaxy s20 FE
        brand: Samsung
        price: 649.0
        screenSize: 6.5
        weight: 0.19
        deviceType: MOBILE
        storage: 128
        processor: Samsung Exynos 990
        setColors: null
        memory: 6
        camera: "Principal 12Mp"
        os: Android
...
    ShopEmpty: 
      type: object
      required: 
        - name
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
    Complement:
      type: array
      items: 
        type: object
        required:
          - name
        properties:
          complementId:
            type: integer
          name: 
            type: string
          material: 
            type: string
          price: 
            type: number
          compatibleDevices:
            type: array
            items: 
              type: object
              required: 
                - name
              properties:
                id:
                  type: integer
                name:
                  type: string
                brand:
                  type: string
                price: 
                  type: number
                screenSize:
                  type: number
                weight:
                  type: number
                deviceType:
                  type: string
                storage:
                  type: integer
                processor:
                  type: string
                memory:
                  type: integer
                OS:
                  type: string
                setColors:
                  type: array
                  items:
                    type: string
                camera:
                  type: string
      example:
        complementId: 25
        name: Silicon Case Iphone 12 Pro Max
        material: Silicon
        price: 50.0
        compatibleDevice:
          id: 1
          name: Galaxy s20 FE
          brand: Samsung
          price: 649.0
          screenSize: 6.5
          weight: 0.19
          deviceType: MOBILE
          storage: 128
          processor: Samsung Exynos 990
          memory: 6
          setColors: [Cloud Orange, Cloud Red, Cloud white, Cloud Navy]
          camera: "Principal 12Mp"
          os: Android
...